requirements.txt
openai==1.55.0
openai-whisper==20231117
python-dotenv==1.0.1
tqdm==4.66.4
